requirements.txt
selenium==4.12.0
webdriver-manager==4.0.1
pytest==7.4.0
allure-pytest==2.13.2
